dashboard/requirements.txt
streamlit==1.26.0
pandas==2.2.3  # Use a precompiled version
google-auth==2.23.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.103.0
serpapi  # Remove version constraint
requests==2.31.0
python-dotenv==1.0.0
